requirements.txt
ultralytics==8.1.2
opencv-python==4.8.1.78
numpy==1.24.3
pillow==10.0.1
torch==2.0.1